week6/HmmSpeakApp/hmmspeak-flask/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai
webrtcvad==2.0.10
pytest==8.0.0
pytest-cov==4.1.0
together